+++ src/routes/movies.js
@@ -1,44 +1,27 @@
+const express = require('express');
+const mongoose = require('mongoose');
 const { Movie, validate } = require('../models/movie');
 const { Genre } = require('../models/genre');
-const express = require('express');
 const router = express.Router();
 
 router.get('/', async (req, res) => {
-    const movies = await Movie.find().sort('name');
-    res.send(movies);
+    try {
+        const movies = await Movie.find().sort('name');
+        res.send(movies);
+    } catch (err) {
+        res.status(500).send(err.message);
+    }
 });
 
 router.post('/', async (req, res) => {
-    const { error } = validate(req.body);
-    if (error) return res.status(400).send(error.details[0].message);
+    try {
+        const { error } = validate(req.body);
+        if (error) return res.status(400).send(error.details[0].message);
 
-    const genre = await Genre.findById(req.body.genreId);
-    if (!genre) return res.status(400).send('Invalid genre.');
+        const genre = await Genre.findById(req.body.genreId);
+        if (!genre) return res.status(400).send('Invalid genre.');
 
-    let movie = new Movie({
-        title: req.body.title,
-        genre: {
-            _id: genre._id,
-            name: genre.name
-        },
-        numberInStock: req.body.numberInStock,
-        dailyRentalRate: req.body.dailyRentalRate
-    });
-    movie = await movie.save();
-
-    res.send(movie);
-});
-
-router.put('/:id', async (req, res) => {
-    const { error } = validate(req.body);
-    if (error) return res.status(400).send(error.details[0].message);
-
-    const genre = await Genre.findById(req.body.genreId);
-    if (!genre) return res.status(400).send('Invalid genre.');
-
-    const movie = await Movie.findByIdAndUpdate(
-        req.params.id,
-        {
+        let movie = new Movie({
             title: req.body.title,
             genre: {
                 _id: genre._id,
@@ -46,38 +29,84 @@ router.put('/:id', async (req, res) => {
             },
             numberInStock: req.body.numberInStock,
             dailyRentalRate: req.body.dailyRentalRate
-        },
-        { new: true }
-    );
+        });
+        movie = await movie.save();
 
-    if (!movie)
-        return res
-            .status(404)
-            .send('The movie with the given ID was not found.');
+        res.send(movie);
+    } catch (err) {
+        res.status(500).send(err.message);
+    }
+});
 
-    res.send(movie);
+router.put('/:id', async (req, res) => {
+    try {
+        const { error } = validate(req.body);
+        if (error) return res.status(400).send(error.details[0].message);
+
+        if (!mongoose.Types.ObjectId.isValid(req.params.id))
+            return res.status(400).send('Invalid Params Id');
+
+        const genre = await Genre.findById(req.body.genreId);
+        if (!genre) return res.status(400).send('Invalid genre.');
+
+        const movie = await Movie.findByIdAndUpdate(
+            req.params.id,
+            {
+                title: req.body.title,
+                genre: {
+                    _id: genre._id,
+                    name: genre.name
+                },
+                numberInStock: req.body.numberInStock,
+                dailyRentalRate: req.body.dailyRentalRate
+            },
+            { new: true }
+        );
+
+        if (!movie)
+            return res
+                .status(404)
+                .send('The movie with the given ID was not found.');
+
+        res.send(movie);
+    } catch (err) {
+        res.status(500).send(err.message);
+    }
 });
 
 router.delete('/:id', async (req, res) => {
-    const movie = await Movie.findByIdAndRemove(req.params.id);
+    try {
+        if (!mongoose.Types.ObjectId.isValid(req.params.id))
+            return res.status(400).send('Invalid Params Id');
 
-    if (!movie)
-        return res
-            .status(404)
-            .send('The movie with the given ID was not found.');
+        const movie = await Movie.findByIdAndRemove(req.params.id);
 
-    res.send(movie);
+        if (!movie)
+            return res
+                .status(404)
+                .send('The movie with the given ID was not found.');
+        res.send(movie);
+    } catch (err) {
+        res.status(500).send(err.message);
+    }
 });
 
 router.get('/:id', async (req, res) => {
-    const movie = await Movie.findById(req.params.id);
+    try {
+        if (!mongoose.Types.ObjectId.isValid(req.params.id))
+            return res.status(400).send('Invalid Params Id');
 
-    if (!movie)
-        return res
-            .status(404)
-            .send('The movie with the given ID was not found.');
+        const movie = await Movie.findById(req.params.id);
 
-    res.send(movie);
+        if (!movie)
+            return res
+                .status(404)
+                .send('The movie with the given ID was not found.');
+
+        res.send(movie);
+    } catch (err) {
+        res.status(500).send(err.message);
+    }
 });
 
 module.exports = router;
